Nodes now recognize when they have been tapped. Doesn't work if the canvas is moved or resized. I'm guessing that those gestures don't update the local position variables of the nodes.

+++ Circl/app/src/main/java/com/calvinpelletier/circl/Node.java
@@ -28,7 +28,7 @@ public class Node {
         this.position = newPosition;
     }
 
-    public void setPosition(int x, int y) {
+    public void setPosition(float x, float y) {
         this.position.x = x;
         this.position.y = y;
     }
